requirements.txt
pyperclip
pycryptodome